Work on Timer

+++ app/src/main/java/com/viam/feeder/data/models/ClockTimer.kt
@@ -2,7 +2,7 @@ package com.viam.feeder.data.models
 
 import com.viam.feeder.R
 
-class ClockTimer(var id: Long = 0, hour: Int, var minute: Int, var time: Int = R.string.am) {
+class ClockTimer(var id: Int = 0, hour: Int, var minute: Int, var time: Int = R.string.am) {
     var hour: Int = hour
         set(value) {
             field = value % 12
